tiktok-scraper/requirements-base.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
supabase>=2.3.0
pandas>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
apscheduler>=3.10.0
psycopg2-binary>=2.9.0
requests>=2.31.0
